// Home/transaction/fin-04-invoice/model/constructionWork.ts
export class ConstructionWork {
    clinicId: number;
    date: Date;
    assetName: string;
    beNumber: string;
    typeOfExpense: string;
    nameOfSupplier: string;
    cwFinalBqCost: number;
    tenderRefNo: string;
    letterOfAwardOrContractRefNo: string;
    certOfPracticalCompletionRefNo: string;
    certOfTcNo: string;
    certOfAcceptanceRef: string;
    agreedTotal: number;

}