// Home/transaction/fin-04-invoice/model/fin-10b.ts
import { ConstructionWork } from './constructionWork';

export class Fin10b {
    stateId: number;
    districtId: number;
    clinicTypeId: number;
    clinicCode: string;
    clinicId: number;
    month: any;
    year: any;
    assetName: string;
    beNumber: string;
    status: string;
    date: any;
    fin10bConstructionWorks: Array<ConstructionWork>;
}